camera/test_camera.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for mozCameras</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="doTest()">

<script class="testbody" type="text/javascript;version=1.7">

function doTest() {
  navigator.mozCameras.getCamera(function() {alert('onsucces')}, function() {alert('onerror')});;
  ok(true, "navigator.mozCameras.getCamera should not throw");

try {
  navigator.mozCameras.getListOfCameras();
  ok(false, "navigator.mozCameras.getListOfCameras should throw, not implemented");
} catch(e) {
  ok(true, "navigator.mozCameras.getListOfCameras should throw, not implemented");
}
SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

</script>
</body>

</html>
